requirements.txt
# Core dependencies
gradio>=5.0.0,<6.0.0
# langchain==0.1.0
# langchain-community==0.0.10
# langchain-groq==0.0.1
# langchain-pinecone==0.0.1
# langchain-core==0.1.10
# langchain-huggingface==0.0.6
langchain==0.3.27
langchain-community==0.3.27
langchain-groq==0.3.7
langchain-pinecone==0.2.11
langchain-huggingface==0.3.1
# Do NOT pin langchain-core explicitly

# LLM and AI
# groq==0.4.2
# huggingface-hub==0.19.4
groq>=0.29.0,<1.0.0
huggingface-hub>=0.24.0,<1.0.0
pinecone>=6.0.0,<7.0.0

# Vector database
# pinecone-client==2.2.4
# pinecone-client>=3.0.0,<4.0.0

# Document processing
PyPDF2==3.0.1
PyMuPDF>=1.25.0,<2.0.0
Pillow>=11.0.0,<12.0.0
pytesseract==0.3.10
python-docx==1.1.0
docx2txt==0.8
beautifulsoup4==4.12.2
email-validator==2.1.0

# NLP and text processing
# spacy==3.7.2
spacy==3.8.7
# nltk==3.8.1
nltk==3.9.1

sentence-transformers==3.0.1

# Enhanced features - Consistency & Interpretability
# (No additional dependencies needed - uses existing libraries)

# Enhanced features - Scalability & Performance
# (No additional dependencies needed - uses existing libraries)

# Enhanced features - Security & Compliance
cryptography==41.0.7

# Enhanced features - Integration Capabilities
flask==3.0.0
flask-cors==4.0.0
requests>=2.32.4

# Environment and configuration
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1
